% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{backtrack}
\alias{backtrack}
\title{Move backwards from the global minimum to a more conservative solution.}
\usage{
backtrack(x, max_total_rise = 0.1, max_step_rise = 0.05)
}
\arguments{
\item{x}{Data.}

\item{max_total_rise}{Value cannot exceed minimum * (1 + max_total_rise).}

\item{max_step_rise}{Each step must be less than minimum * (1 + max_step_rise).}
}
\value{
List of two lists:
\itemize{
\item minimum: $ind
$val
\item backtrack: $ind
$val
}
}
\description{
Move backwards from the global minimum to a more conservative solution.
}
\examples{

ombc_gmm_k3n1000o10 <- ombc_gmm(
  gmm_k3n1000o10[, 1:2],
  comp_num = 3, max_out = 20
)

backtrack(ombc_gmm_k3n1000o10$distrib_diff_mat[, "full"])

}
