% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mahalas.R
\name{compare_mahalas}
\alias{compare_mahalas}
\title{compare_mahalas}
\usage{
compare_mahalas(x, z, mu, Sigma)
}
\arguments{
\item{x}{Data.}

\item{z}{Component assignment probability matrix.}

\item{mu}{List of component mean vectors.}

\item{Sigma}{List of component covariance matrices.}
}
\value{
List of
\itemize{
\item mahala_comparison
\item mahala_comparison_vec
\item choice_id
}
}
\description{
compare_mahalas
}
