% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mahalas.R
\name{compare_mahalas_g}
\alias{compare_mahalas_g}
\title{compare_mahalas_g}
\usage{
compare_mahalas_g(x, z_g, mu_g, Sigma_g)
}
\arguments{
\item{x}{Data.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{Sigma_g}{Component covariance matrix.}
}
\value{
List of
\itemize{
\item mahala_comparison_g
\item scaled_mahalas_g
\item dens_g
}
}
\description{
compare_mahalas_g
}
