% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{count_extremes}
\alias{count_extremes}
\title{distrib_diff_gmm}
\usage{
count_extremes(x, best_model)
}
\arguments{
\item{x}{Data.}

\item{best_model}{mixture::gpcm best_model}
}
\value{
outlier_bool
}
\description{
distrib_diff_gmm
}
