% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{count_extremes_g}
\alias{count_extremes_g}
\title{count_extremes_g}
\usage{
count_extremes_g(x, z_g, mu_g, sigma_g, logdet_g, tail_prop)
}
\arguments{
\item{x}{Data.}

\item{z_g}{Binary version of assignment probability vector for component
g.}

\item{mu_g}{Mean vector for component g.}

\item{sigma_g}{Covariance matrix for component g.}

\item{logdet_g}{Log-determinants of covariance matrix for component g.}

\item{tail_prop}{.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
count_extremes_g
}
