% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_gmm.R
\name{distrib_diff_gmm}
\alias{distrib_diff_gmm}
\title{distrib_diff_gmm}
\usage{
distrib_diff_gmm(x, z, mu, Sigma)
}
\arguments{
\item{x}{Data.}

\item{z}{Component assignment probability matrix.}

\item{mu}{List of component mean vectors.}

\item{Sigma}{List of component covariance matrices.}
}
\value{
List of
\itemize{
\item distrib_diff
\item distrib_diff_vec
\item choice_id
}
}
\description{
distrib_diff_gmm
}
