% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_gmm.R
\name{distrib_diff_gmm_g}
\alias{distrib_diff_gmm_g}
\title{distrib_diff_gmm_g}
\usage{
distrib_diff_gmm_g(x, z_g, mu_g, sigma_g)
}
\arguments{
\item{x}{Data.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{sigma_g}{Component covariance matrix.}
}
\value{
List of
\itemize{
\item distrib_diff_g
\item scaled_mahalas_g
\item dens_g
}
}
\description{
distrib_diff_gmm_g
}
