% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{distrib_diff_lcwm_g}
\alias{distrib_diff_lcwm_g}
\title{distrib_diff_lcwm_g}
\usage{
distrib_diff_lcwm_g(
  x,
  z_g,
  mu_g,
  sigma_g,
  mod_g,
  y_sigma_g,
  dd_weight = 0.5,
  dens_power = 0.5
)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{sigma_g}{Component covariance matrix.}

\item{mod_g}{Component regression model.}

\item{y_sigma_g}{Component regression standard deviation.}

\item{dd_weight}{.}

\item{dens_power}{.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
distrib_diff_lcwm_g
}
