% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_lcwm.R
\name{distrib_diff_lcwm_g_x}
\alias{distrib_diff_lcwm_g_x}
\title{distrib_diff_lcwm_g_x}
\usage{
distrib_diff_lcwm_g_x(x, z_g, mu_g, sigma_g)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{sigma_g}{Component covariance matrix.}
}
\value{
List of
\itemize{
\item distrib_diff_g
\item scaled_mahalas_g
\item dens_g_x
}
}
\description{
distrib_diff_lcwm_g_x
}
