% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_lcwm.R
\name{distrib_diff_lcwm_g_y}
\alias{distrib_diff_lcwm_g_y}
\title{distrib_diff_lcwm_g_y}
\usage{
distrib_diff_lcwm_g_y(x, z_g, mod_g, y_sigma_g)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mod_g}{Component regression model.}

\item{y_sigma_g}{Component regression standard deviation.}
}
\value{
List of
\itemize{
\item distrib_diff_g_y
\item scsqst_res_g
\item dens_g_y
}
}
\description{
distrib_diff_lcwm_g_y
}
