% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_mlr.R
\name{distrib_diff_mlr}
\alias{distrib_diff_mlr}
\title{distrib_diff_mlr}
\usage{
distrib_diff_mlr(student_resids, df)
}
\arguments{
\item{student_resids}{Data.}

\item{df}{Component assignment probability matrix.}
}
\value{
List of
\itemize{
\item distrib_diff
\item sq_st_res
\item choice_id
}
}
\description{
distrib_diff_mlr
}
