% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_elbow.R
\name{find_elbow}
\alias{find_elbow}
\title{Find the elbow using a two-piece linear regression model.}
\usage{
find_elbow(y, search_centre = NULL, concave = TRUE)
}
\arguments{
\item{y}{Numeric vector.}

\item{search_centre}{Centre of changepoint search interval.}

\item{concave}{Logical value.}
}
\value{
List:
\itemize{
\item $choice: a numeric value indicating the elbow's location.
\item $search_interval: a numeric vector of length 2 indicating the minimum and
maximum values considered for the changepoint.
}
}
\description{
Find the elbow using a two-piece linear regression model.
}
