% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gmm.R
\name{fit_gmm}
\alias{fit_gmm}
\title{Fit a Gaussian mixture model given the number of outliers and a
ranking of the candidate outliers.}
\usage{
fit_gmm(
  x,
  outlier_rank,
  outlier_num,
  comp_num,
  mnames = "VVV",
  seed = 123,
  init_z = NULL
)
}
\arguments{
\item{x}{Data set.}

\item{outlier_rank}{Order in which points were trimmed / removed.}

\item{outlier_num}{Number of outliers.}

\item{comp_num}{Number of mixture components.}

\item{mnames}{Gaussian mixture model covariance structure.}

\item{seed}{Seed.}

\item{init_z}{Initial assignment probability matrix.}
}
\value{
List:
\itemize{
\item $labels
\item $mix
}
}
\description{
Fit a Gaussian mixture model given the number of outliers and a
ranking of the candidate outliers.
}
