% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{fixed_ombc_gmm}
\alias{fixed_ombc_gmm}
\title{fixed_ombc_gmm}
\usage{
fixed_ombc_gmm(
  outlier_bool,
  x,
  comp_num,
  max_out,
  mnames = "VVV",
  nmax = 10,
  init_method = c("hc", "kmpp"),
  kmpp_seed = 123,
  print_interval = Inf
)
}
\arguments{
\item{outlier_bool}{Logical vector indicating which observations are outliers
for the fixed model.}

\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Model names for mixture::gpcm.}

\item{nmax}{Maximum number of iterations for mixture::gpcm.}

\item{init_method}{Method used to initialise each mixture model.}

\item{kmpp_seed}{Optional seed for k-means++ initialisation. Default is
hierarchical clustering.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item removal_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
