% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio_mlr.R
\name{idio_mlr}
\alias{idio_mlr}
\title{idio_mlr}
\usage{
idio_mlr(x, y, max_out, print_interval = Inf)
}
\arguments{
\item{x}{Covariate data.}

\item{y}{Response variable.}

\item{max_out}{Maximum number of outliers.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item outlier_bool
\item outlier_num
\item outlier_rank
\item mod
}
}
\description{
Iterative Detection & Identification of Outliers while Clustering
}
\examples{
faithful_plus <- rbind(faithful, c(4.5, 60), c(4.5, 55), c(4.5, 50), c(4.5, 45), c(4.5, 40))
faithful_idio_mlr <- idio_mlr(faithful_plus[, 1], faithful_plus[, 2],
                              max_out = 20)
par(mfrow = c(1, 2))
plot(0:20, faithful_idio_mlr$distrib_diffs, type = "l")
abline(v = faithful_idio_mlr$outlier_num)
plot(faithful_plus, col = 1 + faithful_idio_mlr$outlier_bool)
abline(faithful_idio_mlr$mod)
par(mfrow = c(1, 1))
}
