% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idioClust-package.R
\docType{package}
\name{idioClust-package}
\alias{idioClust}
\alias{idioClust-package}
\title{idioClust: Iterative Detection & Identification of Outliers while Clustering}
\description{
Iterative Detection & Identification of Outliers while Clustering
}
\author{
\strong{Maintainer}: Ultán Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
