% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idioclust.R
\name{idioclust}
\alias{idioclust}
\title{idioclust}
\usage{
idioclust(x, G, max_out, mnames = "VVV", seed = 123, print_progress = FALSE)
}
\arguments{
\item{x}{Data.}

\item{G}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Model names for mixture::gpcm.}

\item{seed}{Seed.}

\item{print_progress}{Should iteration count be printed?}
}
\value{
List of
\itemize{
\item comparisons
\item outlier_bool
\item outlier_num
\item outlier_rank
\item oGMM_labels
}
}
\description{
Iterative Detection & Identification of Outliers while Clustering
}
\examples{
faithful_idio <- idioclust(faithful, G = 2, max_out = 20, seed = 123)
par(mfrow = c(1, 2))
plot(faithful_idio$comparisons, type = "l")
abline(v = faithful_idio$outlier_num)
plot(faithful, col = faithful_idio$oGMM_labels)
par(mfrow = c(1, 1))
}
