% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcwm_k2n1000oxy10}
\alias{lcwm_k2n1000oxy10}
\title{Data set from \code{simulate_lcwm}. 2 components, 1000 true observations,
10 hybrid outliers.}
\format{
\subsection{\code{lcwm_k2n1000oxy10}}{

A data frame with 1010 rows and 3 columns:
\describe{
\item{X1, Y}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1 or 2 for true points.}
}
}
}
\usage{
lcwm_k2n1000oxy10
}
\description{
Simulated data set with 1000 observations from 2 Gaussian components with
linear responses.
}
\keyword{datasets}
