% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcwm_k2n2000oxy50}
\alias{lcwm_k2n2000oxy50}
\title{Data set from \code{simulate_lcwm}. 2 components, 2000 true observations,
50 hybrid outliers.}
\format{
\subsection{\code{lcwm_k2n2000oxy50}}{

A data frame with 2050 rows and 3 columns:
\describe{
\item{X1, Y}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1 or 2 for true points.}
}
}
}
\usage{
lcwm_k2n2000oxy50
}
\description{
Simulated data set with 2000 observations from 2 Gaussian components with
linear responses.
}
\keyword{datasets}
