% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcwm_k3n1000o10}
\alias{lcwm_k3n1000o10}
\title{Data set from \code{simulate_lcwm}. 3 components, 1000 true observations,
10 outliers.}
\format{
\subsection{\code{lcwm_k3n1000o10}}{

A data frame with 1010 rows and 3 columns:
\describe{
\item{X1, Y}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1, 2, or 3 for true points.}
}
}
}
\usage{
lcwm_k3n1000o10
}
\description{
Simulated data set with 1000 observations from 3 Gaussian components and 10
outliers.
}
\keyword{datasets}
