% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcwm_k3n2000o20}
\alias{lcwm_k3n2000o20}
\title{Data set from \code{simulate_lcwm}. 3 components, 2000 true observations,
20 outliers.}
\format{
\subsection{\code{lcwm_k3n2000o20}}{

A data frame with 1010 rows and 3 columns:
\describe{
\item{X1, Y}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1, 2, or 3 for true points.}
}
}
}
\usage{
lcwm_k3n2000o20
}
\description{
Simulated data set with 2000 observations from 3 Gaussian components and 20
outliers.
}
\keyword{datasets}
