% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{new_outliermbc_gmm}
\alias{new_outliermbc_gmm}
\title{Constructor for \code{"outliermbc_gmm"} S3 class.}
\usage{
new_outliermbc_gmm(x = list())
}
\arguments{
\item{x}{List.}
}
\value{
"outliermbc_gmm" S3 object.
}
\description{
Constructor for \code{"outliermbc_gmm"} S3 class.
}
