% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{new_outliermbc_lcwm}
\alias{new_outliermbc_lcwm}
\title{Constructor for "outliermbc_lcwm" S3 object.}
\usage{
new_outliermbc_lcwm(x = list())
}
\arguments{
\item{x}{List.}
}
\value{
"outliermbc_lcwm" S3 object.
}
\description{
Constructor for "outliermbc_lcwm" S3 object.
}
