% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc_gmm}
\alias{ombc_gmm}
\title{ombc_gmm}
\usage{
ombc_gmm(
  x,
  comp_num,
  max_out,
  mnames = "VVV",
  seed = 123,
  reinit_interval = Inf,
  print_interval = Inf,
  gross_outs = NULL
)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Model names for mixture::gpcm.}

\item{seed}{Seed.}

\item{reinit_interval}{How frequently to reinitialise.}

\item{print_interval}{How frequently the iteration count is printed.}

\item{gross_outs}{logical vector identifying gross outliers to be removed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item min_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{
gmm <- simulate_gmm(
  n = c(2000, 1000, 1000),
  mu = list(c(-1, 0), c(+1, -1), c(+1, +1)),
  sigma = list(diag(c(0.2, 4 * 0.2)), diag(c(0.2, 0.2)), diag(c(0.2, 0.2))),
  outlier_num = 40,
  seed = 123,
  crit_val = 0.9999,
  range_multiplier = 1.5
)

ombc_gmm <- ombc_gmm(gmm[, 1:2], comp_num = 3, max_out = 80)

plot(0:80, ombc_gmm$distrib_diffs, type = "l")
abline(v = ombc_gmm$outlier_num)

plot(gmm[, c("X1", "X2")], col = ombc_gmm$labels + 1, pch = gmm$G + 1)
}
