% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{ombc_lcwm}
\alias{ombc_lcwm}
\title{ombc_lcwm}
\usage{
ombc_lcwm(
  xy,
  x,
  y_formula,
  comp_num,
  max_out,
  gross_outs = rep(FALSE, nrow(x)),
  init_scheme = c("update", "reinit", "reuse"),
  mnames = "VVV",
  nmax = 1000,
  atol = 1e-08,
  init_z = NULL,
  init_method = c("hc", "kmpp"),
  kmpp_seed = 123,
  print_interval = Inf,
  alpha = 0.5,
  outlier_type = c("x_and_y", "x_only", "y_only", "hybrid")
)
}
\arguments{
\item{xy}{\code{data.frame} containing covariates and response.}

\item{x}{Covariate data only.}

\item{y_formula}{Regression formula.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{gross_outs}{Logical vector identifying gross outliers.}

\item{init_scheme}{Which initialisation scheme to use.}

\item{mnames}{Model names for mixture::gpcm.}

\item{nmax}{Maximum number of iterations for mixture::gpcm.}

\item{atol}{EM convergence tolerance threshold for mixture::gpcm.}

\item{init_z}{Initial z matrix.}

\item{init_method}{Method used to initialise each mixture model.}

\item{kmpp_seed}{Optional seed for k-means++ initialisation. Default is
hierarchical clustering.}

\item{print_interval}{How frequently the iteration count is printed.}

\item{alpha}{Factor to control relative importance of covariate and response.}

\item{outlier_type}{Combined, covariate, or response outliers.}
}
\value{
List of
*labels = labels,
\itemize{
\item outlier_bool = outlier_bool,
\item outlier_num = outlier_num,
\item outlier_rank = outlier_rank,
\item gross_outs = gross_outs,
\item lcwm = lcwm,
\item loglike = loglike,
\item removal_dens = removal_dens,
\item distrib_diff_vec = distrib_diff_vec,
\item distrib_diff_mat = distrib_diff_mat,
\item call = this_call,
\item version = ombc_version,
\item conv_status = conv_status
}
}
\description{
Iterative Detection & Identification of Outliers for a Linear
Cluster-Weighted Model
}
\examples{

ombc_lcwm_k2n1000oxy10 <- ombc_lcwm(
  xy = lcwm_k2n1000oxy10[, c("X1", "Y")],
  x = lcwm_k2n1000oxy10$X1,
  y_formula = Y ~ X1,
  comp_num = 2,
  max_out = 20,
  mnames = "V"
)

plot_curve(ombc_lcwm_k2n1000oxy10)

}
