% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcast.R
\name{outback_gmm}
\alias{outback_gmm}
\title{Alternate between returning the highest density trimmed point and
refitting a Gaussian mixture model.}
\usage{
outback_gmm(
  x,
  outlier_rank,
  comp_num,
  turning_point = NULL,
  mnames = "VVV",
  seed = 123,
  print_interval = Inf
)
}
\arguments{
\item{x}{Data set.}

\item{outlier_rank}{Order in which points were trimmed / removed.}

\item{comp_num}{Number of mixture components.}

\item{turning_point}{Point at which to start returning points.}

\item{mnames}{Gaussian mixture model covariance structure.}

\item{seed}{Seed.}

\item{print_interval}{Number of iterations between print statements.}
}
\value{
List:
\itemize{
\item $densities
\item $outlier_rank
\item $loglike
}
}
\description{
Alternate between returning the highest density trimmed point and
refitting a Gaussian mixture model.
}
