% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcast.R
\name{outcast_gmm}
\alias{outcast_gmm}
\title{Alternate between trimming the lowest density point and refitting a
Gaussian mixture model.}
\usage{
outcast_gmm(
  x,
  comp_num,
  max_out,
  mnames = "VVV",
  seed = 123,
  reinit_interval = Inf,
  print_interval = Inf
)
}
\arguments{
\item{x}{Data set.}

\item{comp_num}{Number of mixture components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Gaussian mixture model covariance structure.}

\item{seed}{Seed.}

\item{reinit_interval}{Number of iterations between proposed reinitialisations.}

\item{print_interval}{Number of iterations between print statements.}
}
\value{
List:
\itemize{
\item $densities
\item $outlier_rank
\item $loglike
}
}
\description{
Record the order and mixture density of the trimmed points.
}
