% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierMBC-package.R
\docType{package}
\name{outlierMBC-package}
\alias{outlierMBC}
\alias{outlierMBC-package}
\title{outlierMBC: Iterative Detection & Identification of Outliers for Model-Based Clustering}
\description{
Exported Functions:
\itemize{
\item \code{ombc_gmm} - Identify multivariate outliers while clustering the data with
a Gaussian mixture model.
\item \code{ombc_mlr} - Identify response variable outliers while fitting a multiple
linear regression model to the data.
\item \code{simulate_noisy_gmm} - Simulate data from a Gaussian mixture model with
multivariate outliers.
\item \code{simulate_noisy_mlr} - Simulate data from a multiple linear regression
model with response variable outliers.
}
}
\author{
\strong{Maintainer}: UltC!n Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
