% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{plot.outliermbc_gmm}
\alias{plot.outliermbc_gmm}
\title{plot method for \code{"outliermbc_gmm"} S3 class.}
\usage{
\method{plot}{outliermbc_gmm}(x, backtrack = FALSE, ...)
}
\arguments{
\item{x}{List}

\item{backtrack}{Logical}

\item{...}{Other arguments}
}
\value{
A ggplot
}
\description{
plot method for \code{"outliermbc_gmm"} S3 class.
}
