% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{plot.outliermbc_lcwm}
\alias{plot.outliermbc_lcwm}
\title{plot method for \code{"outliermbc_lcwm"} S3 class.}
\usage{
\method{plot}{outliermbc_lcwm}(x, backtrack = FALSE, ...)
}
\arguments{
\item{x}{List}

\item{backtrack}{Logical}

\item{...}{Other arguments}
}
\value{
A ggplot
}
\description{
plot method for \code{"outliermbc_lcwm"} S3 class.
}
