% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{plot_backtrack}
\alias{plot_backtrack}
\title{Plot the dissimilarity curve for the backtrack method.}
\usage{
plot_backtrack(ombc_out, max_total_rise = 0.1, max_step_rise = 0.05)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}

\item{max_total_rise}{Upper limit for the cumulative increase, as a
proportion of the global minimum dissimilarity, from
all backward steps.}

\item{max_step_rise}{Upper limit for the increase, as a proportion of the
global minimum dissimilarity, from each backward step.}
}
\value{
ggplot of the dissimilarity curve showing the minimum solution and
the backtack solutions. The dissimilarity values (y axis) are
divided by the minimum dissimilarity so that the rescaled minimum is
at 1.
}
\description{
Plot the dissimilarity curve for the backtrack method.
}
