% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{plot_backtrack_curve}
\alias{plot_backtrack_curve}
\title{Plot the outlier number selection curve for the backtrack method.}
\usage{
plot_backtrack_curve(ombc_out, max_value, max_step)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}

\item{max_value}{Value cannot exceed minimum * (1 + max_value).}

\item{max_step}{Each step must be less than minimum * (1 + max_step).}
}
\value{
A gg object.
}
\description{
Plot the outlier number selection curve for the backtrack method.
}
