% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot multiple dissimilarity curves.}
\usage{
plot_comparison(ombc_list)
}
\arguments{
\item{ombc_list}{A list of outputs from ombc_gmm.}
}
\value{
A gg object.
}
\description{
Plot multiple dissimilarity curves.
}
