% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_comparison_bic}
\alias{plot_comparison_bic}
\title{Plot multiple dissimilarity curves.}
\usage{
plot_comparison_bic(ombc_list)
}
\arguments{
\item{ombc_list}{A list of outputs from \code{ombc_gmm}.}
}
\value{
\code{plot_comparison} returns a ggplot object consisting of multiple
dissimilarity curves overlaid on the same plot.
}
\description{
Given a range of \link{ombc_gmm} outputs, each arising from a different model,
this function is designed to produce a graphical aid for selecting the best
model. It displays the dissimilarity curves from each of these models on the
same plot.
}
