% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_full_curve}
\alias{plot_full_curve}
\title{Plot the outlier number selection curve for the full method.}
\usage{
plot_full_curve(ombc_out)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}
}
\value{
A gg object.
}
\description{
Plot the outlier number selection curve for the full method.
}
