% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gross.R
\name{plot_gross}
\alias{plot_gross}
\title{Plot kNN distance values to determine the \code{search_centre} for
\code{find_gross}.}
\usage{
plot_gross(x, k_neighbours = floor(nrow(x)/100))
}
\arguments{
\item{x}{Data.}

\item{k_neighbours}{Number of neighbours for dbscan::kNNdist.}
}
\value{
A ggplot2 object
}
\description{
Plot kNN distance values to determine the \code{search_centre} for
\code{find_gross}.
}
