% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_ks_curve}
\alias{plot_ks_curve}
\title{Plot the outlier number selection curve for the ks method.}
\usage{
plot_ks_curve(ombc_out, alpha = 0.05)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}

\item{alpha}{.}
}
\value{
A gg object.
}
\description{
Plot the outlier number selection curve for the ks method.
}
