% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_lcwm_dual}
\alias{plot_lcwm_dual}
\title{Plot three dissimilarity curves for \code{outlierMBC-LCWM}.}
\usage{
plot_lcwm_dual(ombc_out)
}
\arguments{
\item{ombc_out}{An \code{"outliermbc_lcwm"} object, i.e. an output from
\code{ombc_lcwm}.}
}
\value{
\code{plot_lcwm_dual} returns a ggplot object showing three dissimilarity curves
and marking the minimum solution with a vertical line.
}
\description{
Given the output from \link{ombc_lcwm}, this function extracts three dissimilarity
values associated with each outlier number and plots them as a
curve. These represent the actual overall aggregated dissimilarity, and two
aggregated dissimilarities corresponding to the explanatory variables and the
response variable, respectively. It also draws a vertical line at the outlier
number which minimised the overall aggregated dissimilarity.
}
