% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.outliermbc_lcwm}
\alias{print.outliermbc_lcwm}
\title{print method for \code{"outliermbc_lcwm"} S3 class.}
\usage{
\method{print}{outliermbc_lcwm}(x, backtrack = FALSE, max_total_rise = 0.1, max_step_rise = 0.05, ...)
}
\arguments{
\item{x}{List}

\item{backtrack}{Logical}

\item{max_total_rise}{Upper limit for the cumulative increase, as a
proportion of the global minimum dissimilarity, from
all backward steps.}

\item{max_step_rise}{Upper limit for the increase, as a proportion of the
global minimum dissimilarity, from each backward step.}

\item{...}{Other arguments}
}
\value{
A ggplot
}
\description{
print method for \code{"outliermbc_lcwm"} S3 class.
}
