% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_noisy_gmm.R
\name{simulate_noisy_gmm}
\alias{simulate_noisy_gmm}
\title{simulate_noisy_gmm}
\usage{
simulate_noisy_gmm(
  n,
  mu,
  Sigma,
  outlier_num,
  seed = 123,
  crit_val = 0.999,
  unif_range_multiplier = 1.5
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{Sigma}{List of component covariance matrices.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{unif_range_multiplier}{How much greater should the range of the Uniform
samples be than the range of the Normal samples?}
}
\value{
Matrix with a label column.
}
\description{
simulate_noisy_gmm
}
\examples{
n_vec <- c(100, 100)
mu_list <- list(c(-1, -1), c(1, 1))
Sigma_list <- list(diag(c(0.1, 0.1)), diag(c(0.1, 0.1)))
noisy_gmm <- simulate_noisy_gmm(n_vec, mu_list, Sigma_list, 123, 20)
plot(noisy_gmm[, 1:2], col = 1 + noisy_gmm[, 3])
}
