% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_noisy_lcwm.R
\name{simulate_noisy_lcwm}
\alias{simulate_noisy_lcwm}
\title{simulate_noisy_lcwm}
\usage{
simulate_noisy_lcwm(
  n,
  mu,
  sigma,
  beta,
  error_sd,
  outlier_num,
  seed = 123,
  crit_val = 0.9999,
  unif_range_multipliers = c(1.5, 1.5),
  print_interval = Inf
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{unif_range_multipliers}{.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
Matrix with a label column.
}
\description{
Simulate a multiple linear regression model with response variable outliers.
}
\examples{
noisy_lcwm_p1 <- simulate_noisy_lcwm(
  n = c(1000, 1000),
  mu = list(c(-1), c(+1)),
  sigma = list(as.matrix(0.1), as.matrix(0.1)),
  beta = list(c(-1, 1), c(1, 1)),
  error_sd = c(0.5, 0.5),
  outlier_num = c(10, 10),
  seed = 123,
  crit_val = 0.9999
)
plot(
  x = noisy_lcwm_p1$covariates[, 1], y = noisy_lcwm_p1$responses,
  col = 1 + noisy_lcwm_p1$labels, pch = 1 + noisy_lcwm_p1$labels
)
}
