% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_noisy_mlr.R
\name{simulate_noisy_mlr}
\alias{simulate_noisy_mlr}
\title{simulate_noisy_mlr}
\usage{
simulate_noisy_mlr(
  n,
  mu,
  sigma,
  beta,
  error_sd,
  outlier_num,
  seed = 123,
  crit_val = 0.9999,
  unif_range_multiplier = 1.5,
  print_interval = Inf
)
}
\arguments{
\item{n}{Number of true observations.}

\item{mu}{Covariate mean vector.}

\item{sigma}{Covariate covariance matrix.}

\item{beta}{Regression coefficient vector.}

\item{error_sd}{Regression error standard deviations.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{unif_range_multiplier}{How much greater should the range of the Uniform
samples be than the range of the Normal samples?}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List:
\itemize{
\item covariates: matrix of samples from a multivariate Normal distribution.
\item responses: vector of regression dependent variable values.
\item labels: vector of 1s (true observations) and 0s (outliers).
}
}
\description{
Simulate a multiple linear regression model with response variable outliers.
}
\examples{
n <- 1000
mu <- c(+1)
sigma <- as.matrix(0.1)
beta <- c(1, 1)
error_sd <- 0.5
noisy_mlr_p1 <- simulate_noisy_mlr(
  n, mu, sigma, beta,
  error_sd,
  outlier_num = 20, seed = 123,
  crit_val = 0.9999
)
plot(
  x = noisy_mlr_p1$covariates[, 1], y = noisy_mlr_p1$responses,
  col = 1 + noisy_mlr_p1$labels, pch = 1 + noisy_mlr_p1$labels
)
}
