% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{try_mixture_gpcm}
\alias{try_mixture_gpcm}
\title{Run \code{mixture::gpcm} and try alternative covariance structures or
initialisations if necessary.}
\usage{
try_mixture_gpcm(x, comp_num, mnames, z, nmax, atol)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of mixture components.}

\item{mnames}{Model names for mixture::gpcm.}

\item{z}{Component assignment probability matrix for initialisation.}

\item{nmax}{Maximum number of iterations for \code{mixture::gpcm}.}

\item{atol}{EM convergence tolerance threshold for \code{mixture::gpcm}.}
}
\value{
Object of class \code{"gpcm"} outputted by \code{mixture::gpcm}.
}
\description{
If \code{mixture::gpcm} returns an error, this function first tries
the other covariance structures, and then tries a k-means initialisation.
}
