% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lcwm.R
\name{uniform_sample_lcwm}
\alias{uniform_sample_lcwm}
\title{Sample a potential outlier.}
\usage{
uniform_sample_lcwm(
  outlier_type,
  mu_g,
  sigma_g,
  beta_g,
  error_sd_g,
  uniform_spans_g
)
}
\arguments{
\item{outlier_type}{.}

\item{mu_g}{Covariate mean vector for one component.}

\item{sigma_g}{Covariate covariance matrix for one component.}

\item{beta_g}{Regression coefficient vector for one component.}

\item{error_sd_g}{Regression error standard deviation for one component.}

\item{uniform_spans_g}{Covariate and response error ranges for one component.}
}
\value{
List:
\itemize{
\item x = Covariate sample
\item y = Response sample
}
}
\description{
Sample a potential outlier.
}
