% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lcwm.R
\name{uniform_spans_lcwm}
\alias{uniform_spans_lcwm}
\title{Obtain the span of the observations for each component.}
\usage{
uniform_spans_lcwm(range_multipliers, covariates_g, errors_g)
}
\arguments{
\item{range_multipliers}{The sampling region for the Uniform distribution used to simulate proposed
outliers is controlled by multiplying the component widths by these values.}

\item{covariates_g}{Covariate values of the sampled observations.}

\item{errors_g}{Response errors of the sampled observations.}
}
\value{
\code{uniform_spans_lcwm} returns a 2-column matrix. The final row contains the
minimum and maximum values of the response errors, while the previous rows
contain the minimum and maximum values for each covariate.
}
\description{
Determine the minimum and maximum values for each covariate / explanatory
variable and for the response errors from all Gaussian observations.
}
