% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{validate_outliermbc_gmm}
\alias{validate_outliermbc_gmm}
\title{Validator for \code{"outliermbc_gmm"} S3 class.}
\usage{
validate_outliermbc_gmm(x)
}
\arguments{
\item{x}{List.}
}
\description{
Validator for \code{"outliermbc_gmm"} S3 class.
}
