% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{backtrack_gmm}
\alias{backtrack_gmm}
\title{Title}
\usage{
backtrack_gmm(
  x,
  ombc_out,
  max_value = 0.1,
  max_step = 0.01,
  init_model = NULL,
  init_z = NULL
)
}
\arguments{
\item{x}{Data.}

\item{ombc_out}{Output from ombc_gmm.}

\item{max_value}{Value cannot exceed minimum * (1 + max_value).}

\item{max_step}{Each step must be less than minimum * (1 + max_step).}

\item{init_model}{mixture::gpcm best_model}

\item{init_z}{Initial z matrix.}
}
\value{
List:
\itemize{
\item labels
\item outlier_bool
\item outlier_num
\item mix
}
}
\description{
Title
}
\examples{

ombc_gmm_k3n1000o10 <- ombc_gmm(
  gmm_k3n1000o10[, 1:2],
  comp_num = 3, max_out = 20
)

backtrack_gmm(gmm_k3n1000o10[, 1:2], ombc_gmm_k3n1000o10, 0.1, 0.01)
}
