% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{backtrack_lcwm}
\alias{backtrack_lcwm}
\title{Fit a LCWM to the backtrack solution.}
\usage{
backtrack_lcwm(
  xy,
  x,
  ombc_lcwm_out,
  max_total_rise = 0.1,
  max_step_rise = 0.05,
  init_z = NULL
)
}
\arguments{
\item{xy}{\code{data.frame} containing covariates and response.}

\item{x}{Covariate data only.}

\item{ombc_lcwm_out}{Output from ombc_lcwm.}

\item{max_total_rise}{Value cannot exceed minimum * (1 + max_total_rise).}

\item{max_step_rise}{Each step must be less than minimum * (1 + max_step_rise).}

\item{init_z}{Initial z matrix.}
}
\value{
List:
\itemize{
\item labels
\item outlier_bool
\item outlier_num
\item lcwm
\item call
}
}
\description{
Fit a LCWM to the backtrack solution.
}
\examples{

ombc_lcwm_k2n1000oxy10 <- ombc_lcwm(
  xy = lcwm_k2n1000oxy10[, c("X1", "Y")],
  x = lcwm_k2n1000oxy10$X1,
  y_formula = Y ~ X1,
  comp_num = 2,
  max_out = 20,
  mnames = "V"
)

backtrack_lcwm_k2n1000oxy10 <- backtrack_lcwm(
  xy = lcwm_k2n1000oxy10[, c("X1", "Y")],
  x = lcwm_k2n1000oxy10$X1,
  ombc_lcwm_out = ombc_lcwm_k2n1000oxy10
)
}
