% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{count_extremes}
\alias{count_extremes}
\title{distrib_diff_gmm}
\usage{
count_extremes(x, bin_z, prop, mu, sigma, logdet, dens)
}
\arguments{
\item{x}{Data.}

\item{bin_z}{Binary version of component assignment probability matrix.}

\item{prop}{Vector of component proportions.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{logdet}{Vector of log-determinants for covariance matrices.}

\item{dens}{.}
}
\value{
List of
\itemize{
\item distrib_diff
\item distrib_diff_vec
\item choice_id
\item removal_dens
}
}
\description{
distrib_diff_gmm
}
