% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{distrib_diff_gmm}
\alias{distrib_diff_gmm}
\title{distrib_diff_gmm}
\usage{
distrib_diff_gmm(x, z, prop, mu, sigma, logdet, p_range = c(1, 2))
}
\arguments{
\item{x}{Data.}

\item{z}{Component assignment probability matrix.}

\item{prop}{Vector of component proportions.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{logdet}{Vector of log-determinants for covariance matrices.}

\item{p_range}{Range for power mean parameter, p, when summarising CDF
differences.}
}
\value{
List of
\itemize{
\item distrib_diff
\item distrib_diff_vec
\item choice_id
\item removal_dens
}
}
\description{
distrib_diff_gmm
}
