% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{distrib_diff_lcwm}
\alias{distrib_diff_lcwm}
\title{distrib_diff_lcwm}
\usage{
distrib_diff_lcwm(
  x,
  z,
  prop,
  mu,
  sigma,
  mod_list,
  y_sigma,
  dd_weight = 0.5,
  dens_power = 0.5
)
}
\arguments{
\item{x}{Covariate data only.}

\item{z}{Component assignment probability matrix.}

\item{prop}{Vector of component proportions.}

\item{mu}{Matrix of component mean vectors.}

\item{sigma}{Array of component covariance matrices.}

\item{mod_list}{List of component regression models.}

\item{y_sigma}{Vector of component regression standard deviations.}

\item{dd_weight}{.}

\item{dens_power}{.}
}
\value{
List of
\itemize{
\item distrib_diff
\item distrib_diff_vec
\item choice_id
}
}
\description{
distrib_diff_lcwm
}
