% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_lcwm.R
\name{distrib_diff_lcwm}
\alias{distrib_diff_lcwm}
\title{distrib_diff_lcwm}
\usage{
distrib_diff_lcwm(
  x,
  z,
  prop,
  mu,
  sigma,
  mod_list,
  y_sigma,
  alpha = 0.5,
  outlier_type = c("x_and_y", "x_only", "y_only")
)
}
\arguments{
\item{x}{Covariate data only.}

\item{z}{Component assignment probability matrix.}

\item{prop}{Vector of component proportions.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{mod_list}{List of component regression models.}

\item{y_sigma}{Vector of component regression standard deviations.}

\item{alpha}{Factor to control relative importance of covariate and response.}

\item{outlier_type}{Combined, covariate, or response outliers.}
}
\value{
List of
\itemize{
\item distrib_diff
\item distrib_diff_vec
\item choice_id
}
}
\description{
distrib_diff_lcwm
}
