% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{distrib_diff_lcwm_g}
\alias{distrib_diff_lcwm_g}
\title{distrib_diff_lcwm_g}
\usage{
distrib_diff_lcwm_g(
  x,
  z_g,
  mu_g,
  sigma_g,
  mod_g,
  y_sigma_g,
  alpha = 0.5,
  outlier_type = c("x_and_y", "x_only", "y_only", "hybrid")
)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{sigma_g}{Component covariance matrix.}

\item{mod_g}{Component regression model.}

\item{y_sigma_g}{Component regression standard deviation.}

\item{alpha}{Factor to control relative importance of covariate and response.}

\item{outlier_type}{Combined, covariate, or response outliers.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
distrib_diff_lcwm_g
}
