% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{distrib_diff_mahalanobis}
\alias{distrib_diff_mahalanobis}
\title{distrib_diff_mahalanobis}
\usage{
distrib_diff_mahalanobis(x, z_g, mu_g, sigma_g, logdet_g, bool_g)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mu_g}{Component mean vector.}

\item{sigma_g}{Component covariance matrix.}

\item{logdet_g}{Log-determinants for covariance matrix of component g.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
distrib_diff_mahalanobis
}
