% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{distrib_diff_mahalanobis}
\alias{distrib_diff_mahalanobis}
\title{distrib_diff_mahalanobis}
\usage{
distrib_diff_mahalanobis(x, z_g, mu_g, sigma_g, logdet_g, p_range = c(1, 2))
}
\arguments{
\item{x}{Data.}

\item{z_g}{Assignment probability vector for component g.}

\item{mu_g}{Mean vector for component g.}

\item{sigma_g}{Covariance matrix for component g.}

\item{logdet_g}{Log-determinants of covariance matrix for component g.}

\item{p_range}{Range for power mean parameter, p, when summarising CDF
differences.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
distrib_diff_mahalanobis
}
