% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff_mlr.R
\name{distrib_diff_mlr}
\alias{distrib_diff_mlr}
\title{distrib_diff_mlr}
\usage{
distrib_diff_mlr(lm_mod)
}
\arguments{
\item{lm_mod}{Output from \code{lm} function.}
}
\value{
List of
\itemize{
\item distrib_diff
\item scsqst_res
\item choice_id
}
}
\description{
distrib_diff_mlr
}
