% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_diff.R
\name{distrib_diff_residual}
\alias{distrib_diff_residual}
\title{Compute the response dissimilarity for a single component of a Linear
CWM.}
\usage{
distrib_diff_residual(x, z_g, mod_g, y_sigma_g)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mod_g}{Component regression model.}

\item{y_sigma_g}{Component regression standard deviation for the response.}
}
\value{
\code{distrib_diff_lcwm_residual} returns a list with the following elements:
\describe{
\item{\code{diff}}{Response dissimilarity value for this component.}
\item{\code{dens}}{Response density of all observations for this component.}
}
}
\description{
Computes the response dissimilarity value and the response density for every
observation.
}
