% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{distrib_diff_residual}
\alias{distrib_diff_residual}
\title{distrib_diff_residual}
\usage{
distrib_diff_residual(x, z_g, mod_g, y_sigma_g)
}
\arguments{
\item{x}{Covariate data only.}

\item{z_g}{Component assignment probability vector.}

\item{mod_g}{Component regression model.}

\item{y_sigma_g}{Component regression standard deviation.}
}
\value{
List of
\itemize{
\item diff
\item dens
}
}
\description{
distrib_diff_residual
}
