% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gross.R
\name{find_gross}
\alias{find_gross}
\title{Find the gross outliers.}
\usage{
find_gross(
  x,
  max_out,
  multiplier = 3,
  k_neighbours = floor(nrow(x)/100),
  manual_gross_threshold = NULL,
  scale = TRUE
)
}
\arguments{
\item{x}{Data.}

\item{max_out}{Maximum number of outliers.}

\item{multiplier}{Factor by which the \code{max_out}th kNN Distance is multiplied
to get the gross outlier threshold.}

\item{k_neighbours}{Number of neighbours for dbscan::kNNdist.}

\item{manual_gross_threshold}{Optional preset number of gross outliers.}

\item{scale}{Logical}
}
\value{
List:
\itemize{
\item $choice: a numeric value indicating the elbow's location.
\item $bool: a logical vector identifying the gross outliers.
\item $plot
}
}
\description{
Find the gross outliers.
}
