% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gross.R
\name{find_gross}
\alias{find_gross}
\title{Find the gross outliers.}
\usage{
find_gross(x, max_out, search_centre, k_neighbours = 10, choice = NULL)
}
\arguments{
\item{x}{Data set.}

\item{max_out}{Maximum number of outliers.}

\item{search_centre}{Centre of changepoint search interval.}

\item{k_neighbours}{Number of neighbours for dbscan::glosh}

\item{choice}{Optional preset number of gross outliers.}
}
\value{
List:
\itemize{
\item $choice: a numeric value indicating the elbow's location.
\item $bool: a logical vector identifying the gross outliers.
\item $plot
}
}
\description{
Find the gross outliers.
}
