% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gross.R
\name{find_gross}
\alias{find_gross}
\title{Find the gross outliers.}
\usage{
find_gross(
  x,
  max_out,
  k_neighbours = floor(nrow(x)/100),
  underestimate = max(0.6, min(0.8, 0.9 - (30/elbow_choice))),
  search_centre = NULL,
  choice = NULL
)
}
\arguments{
\item{x}{Data.}

\item{max_out}{Maximum number of outliers.}

\item{k_neighbours}{Number of neighbours for dbscan::kNNdist.}

\item{underestimate}{Factor by which to multiply the elbow estimate of the
number of outliers to get the number of gross outliers.}

\item{search_centre}{Optional centre of elbow search interval.}

\item{choice}{Optional preset number of gross outliers.}
}
\value{
List:
\itemize{
\item $choice: a numeric value indicating the elbow's location.
\item $bool: a logical vector identifying the gross outliers.
\item $plot
}
}
\description{
Find the gross outliers.
}
