% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmm_k3n2000o20}
\alias{gmm_k3n2000o20}
\title{Simulated data set consisting of 2000 observations from 3 Gaussian
components and 20 outliers.}
\format{
\subsection{\code{gmm_k3n2000o20}}{

A data frame with 2020 rows and 3 columns:
\describe{
\item{X1, X2}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1, 2, or 3 for true points.}
}
}
}
\source{
For simulation code, see \code{gmm_k3n2000o20.R} in \code{data-raw} folder at
\url{https://github.com/UltanPDoherty/outlierMBC}.
}
\usage{
gmm_k3n2000o20
}
\description{
This data set was simulated using \code{simulate_gmm}. There are 1000 observations
in Component 1, 500 observations in Component 2, and 500 observations in
Component 3.
}
\keyword{datasets}
