% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmm_k3n4000o40}
\alias{gmm_k3n4000o40}
\title{Simulated data set consisting of 4000 observations from 3 Gaussian
components and 40 outliers.}
\format{
\subsection{\code{gmm_k3n4000o40}}{

A data frame with 4040 rows and 3 columns:
\describe{
\item{X1, X2}{Continuous variables.}
\item{G}{Component label: 0 for outliers; 1, 2, or 3 for true points.}
}
}
}
\source{
\url{https://github.com/UltanPDoherty/outlierMBC/blob/main/data-raw/gmm_k3n4000o40.R}
}
\usage{
gmm_k3n4000o40
}
\description{
This data set was simulated using \code{simulate_gmm}. There are 2000 observations
in Component 1, 1000 observations in Component 2, and 1000 observations in
Component 3.
}
\keyword{datasets}
