% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio_mlr.R
\name{idio_mlr}
\alias{idio_mlr}
\title{idio_mlr}
\usage{
idio_mlr(x, y, max_out, print_interval = Inf)
}
\arguments{
\item{x}{Covariate data.}

\item{y}{Response variable.}

\item{max_out}{Maximum number of outliers.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item outlier_bool
\item outlier_num
\item outlier_rank
\item mod
}
}
\description{
Iterative Detection & Identification of Outliers for Multiple Linear
Regression
}
\examples{
n_vec <- c(1000)
mu_list <- list(+1)
sigma_list <- list(as.matrix(0.1))
beta_list <- list(c(1, 1))
error_sd_vec <- c(0.5)
noisy_mlr_p1 <- simulate_noisy_mlr(n_vec, mu_list, sigma_list, beta_list,
                                   error_sd_vec,
                                   outlier_num = 20, seed = 123,
                                   crit_val = 0.9999)
idio_mlr_p1 <- idio_mlr(noisy_mlr_p1$covariates, noisy_mlr_p1$responses,
                        max_out = 40)
# par(mfrow = c(1, 2))
# plot(0:40, idio_mlr_p1$distrib_diffs, type = "l")
# abline(v = idio_mlr_p1$outlier_num)
# plot(x = noisy_mlr_p1$covariates[, 1], y = noisy_mlr_p1$responses,
#      pch = 1 + noisy_mlr_p1$labels, col = 1 + idio_mlr_p1$outlier_bool)
# par(mfrow = c(1, 1))
}
