% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idioClust-package.R
\docType{package}
\name{idioClust-package}
\alias{idioClust}
\alias{idioClust-package}
\title{idioClust: Iterative Detection & Identification of Outliers while Clustering}
\description{
Exported Functions:
\itemize{
\item \code{idio_gmm} - Identify multivariate outliers while clustering the data with a
Gaussian mixture model.
\item \code{idio_mlr} - Identify response variable outliers while fitting a multiple
linear regression model to the data.
\item \code{simulate_noisy_gmm} - Simulate data from a Gaussian mixture model with
multivariate outliers.
\item \code{simulate_noisy_mlr} - Simulate data from a multiple linear regression model
with response variable outliers.
}
}
\author{
\strong{Maintainer}: Ultán Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
