% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc2_gmm}
\alias{ombc2_gmm}
\title{ombc_gmm}
\usage{
ombc2_gmm(x, ombc1, stabilisation_point = 1)
}
\arguments{
\item{x}{Data.}

\item{ombc1}{Output from \code{ombc1_gmm1}.}

\item{stabilisation_point}{Number of points to be trimmed, chosen based on ombc1.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item removal_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{

ombc1_gmm_k3n1000o10 <- ombc1_gmm(
  gmm_k3n1000o10[, 1:2],
  comp_num = 3, max_out = 20
)

ombc2_gmm_k3n1000o10 <- ombc2_gmm(
  gmm_k3n1000o10[, 1:2],
  ombc1_gmm_k3n1000o10,
  stabilisation_point = 1
)

ombc2_gmm_k3n1000o10$plot_curves
ombc2_gmm_k3n1000o10$plot_removal

}
