% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc_gmm}
\alias{ombc_gmm}
\title{ombc_gmm}
\usage{
ombc_gmm(
  x,
  comp_num,
  max_out,
  gross_outs = NULL,
  p_range = c(1, 2),
  mnames = "VVV",
  print_interval = Inf
)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{gross_outs}{logical vector identifying gross outliers to be removed.}

\item{p_range}{Range for power mean parameter, p, when summarising CDF
differences.}

\item{mnames}{Model names for mixture::gpcm.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item min_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{

ombc_gmm_k3n1000o10 <- ombc_gmm(
  gmm_k3n1000o10[, 1:2], comp_num = 3, max_out = 40
)

ombc_gmm_k3n1000o10$plot_curves

ombc_gmm_k3n1000o10$plot_choice

}
