% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc_gmm}
\alias{ombc_gmm}
\title{ombc_gmm}
\usage{
ombc_gmm(
  x,
  comp_num,
  max_out,
  mnames = "VVV",
  seed = 123,
  print_interval = Inf
)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Model names for mixture::gpcm.}

\item{seed}{Seed.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item outlier_bool
\item outlier_num
\item outlier_rank
\item gmm_labels
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{
n_vec <- c(2000, 1000, 1000)
mu_list <- list(c(-1, 0), c(+1, -1), c(+1, +1))
sigma_list <- list(
  diag(c(0.2, 4 * 0.2)),
  diag(c(0.2, 0.2)),
  diag(c(0.2, 0.2))
)
noisy_gmm_p2g3 <- simulate_noisy_gmm(
  n_vec, mu_list, sigma_list,
  outlier_num = 40, seed = 123, crit_val = 0.9999,
  unif_range_multiplier = 1.5
)
ombc_gmm_p2g3 <- ombc_gmm(noisy_gmm_p2g3[, 1:2], comp_num = 3, max_out = 80)
# par(mfrow = c(1, 2))
# plot(0:80, ombc_gmm_p2g3$distrib_diffs, type = "l")
# abline(v = ombc_gmm_p2g3$outlier_num)
# plot(noisy_gmm_p2g3[, 1:2], col = ombc_gmm_p2g3$gmm_labels,
#      pch = 1 + noisy_gmm_p2g3[, 3])
# par(mfrow = c(1, 1))
}
