% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc_gmm}
\alias{ombc_gmm}
\title{ombc_gmm}
\usage{
ombc_gmm(
  x,
  comp_num,
  max_out,
  gross_outs = rep(FALSE, nrow(x)),
  resets = c(),
  mnames = "VVV",
  nmax = 1000,
  atol = 1e-08,
  init_z = NULL,
  init_model = NULL,
  init_method = c("hc", "kmpp"),
  kmpp_seed = 123,
  print_interval = Inf
)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{gross_outs}{Logical vector identifying gross outliers.}

\item{resets}{Vector of iteration numbers at which to reinitialise.}

\item{mnames}{Model names for mixture::gpcm.}

\item{nmax}{Maximum number of iterations for mixture::gpcm.}

\item{atol}{EM convergence tolerance threshold for mixture::gpcm.}

\item{init_z}{Initial z matrix.}

\item{init_model}{mixture::gpcm best_model}

\item{init_method}{Method used to initialise each mixture model.}

\item{kmpp_seed}{Optional seed for k-means++ initialisation. Default is
hierarchical clustering.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item removal_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{

ombc_gmm_k3n1000o10 <- ombc_gmm(
  gmm_k3n1000o10[, 1:2],
  comp_num = 3, max_out = 20
)

ombc_gmm_k3n1000o10$plot_tail_curve
ombc_gmm_k3n1000o10$plot_full_curve

}
