% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_gmm.R
\name{ombc_gmm}
\alias{ombc_gmm}
\title{ombc_gmm}
\usage{
ombc_gmm(
  x,
  comp_num,
  max_out,
  gross_outs = rep(FALSE, nrow(x)),
  expect_num = c(1, 10, 100),
  accept_num = expect_num + c(1, 2, 4),
  reject_num = accept_num + c(2, 4, 8),
  mnames = "VVV",
  nmax = 10,
  print_interval = Inf
)
}
\arguments{
\item{x}{Data.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{gross_outs}{Logical vector identifying gross outliers.}

\item{expect_num}{.}

\item{accept_num}{.}

\item{reject_num}{.}

\item{mnames}{Model names for mixture::gpcm.}

\item{nmax}{Maximum number of iterations for mixture::gpcm.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item distrib_diff_mat
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_gmm
\item loglike
\item removal_dens
}
}
\description{
Iterative Detection & Identification of Outliers for a Gaussian Mixture Model
}
\examples{

ombc_gmm_k3n1000o10 <- ombc_gmm(
  gmm_k3n1000o10[, 1:2],
  comp_num = 3, max_out = 20
)

ombc_gmm_k3n1000o10$plot_curves
ombc_gmm_k3n1000o10$plot_removal

}
