% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ombc_lcwm.R
\name{ombc_lcwm}
\alias{ombc_lcwm}
\title{ombc_lcwm}
\usage{
ombc_lcwm(
  xy,
  x,
  y_formula,
  comp_num,
  max_out,
  mnames = "VVV",
  seed = 123,
  print_interval = Inf,
  alpha = 0.5,
  outlier_type = c("x_and_y", "x_only", "y_only", "hybrid")
)
}
\arguments{
\item{xy}{\code{data.frame} containing covariates and response.}

\item{x}{Covariate data only.}

\item{y_formula}{Regression formula.}

\item{comp_num}{Number of components.}

\item{max_out}{Maximum number of outliers.}

\item{mnames}{Model names for flexCWM::cwm.}

\item{seed}{Seed.}

\item{print_interval}{How frequently the iteration count is printed.}

\item{alpha}{Factor to control relative importance of covariate and response.}

\item{outlier_type}{Combined, covariate, or response outliers.}
}
\value{
List of
\itemize{
\item distrib_diffs
\item outlier_bool
\item outlier_num
\item outlier_rank
\item labels
\item final_cwm
}
}
\description{
Iterative Detection & Identification of Outliers for a Linear
Cluster-Weighted Model
}
\examples{
lcwm <- simulate_lcwm(
  n = c(1000, 1000),
  mu = list(c(-1), c(+1)),
  sigma = list(as.matrix(0.2), as.matrix(0.2)),
  beta = list(c(1, 0), c(1, 3)),
  error_sd = c(1, 1),
  outlier_num = c(25, 25),
  outlier_type = "x_and_y",
  seed = 123,
  crit_val = 0.9999,
  range_multipliers = c(1.5, 2)
)

ombc_lcwm <- ombc_lcwm(
  xy = lcwm[, c("X1", "Y")],
  x = lcwm$X1,
  y_formula = Y ~ X1,
  comp_num = 2,
  max_out = 100,
  mnames = "V",
  seed = 123
)

plot(0:100, ombc_lcwm$distrib_diffs, type = "l")
abline(v = ombc_lcwm$outlier_num)

plot(lcwm[, c("X1", "Y")], col = ombc_lcwm$labels + 1, pch = lcwm$G + 1)
}
