% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierMBC-package.R
\docType{package}
\name{outlierMBC-package}
\alias{outlierMBC}
\alias{outlierMBC-package}
\title{outlierMBC: Model-Based Clustering with Sequential Outlier Identification using the Distribution of Mahalanobis Distances}
\description{
Sequential outlier identification for Gaussian mixture models and linear cluster-weighted models. The optimal number of outliers is chosen based on the dissimilarity between the theoretical and observed distributions of the scaled squared sample Mahalanobis distances. Doherty, McNicholas, and White (2025) \doi{10.48550/arXiv.2505.11668}.
}
\author{
\strong{Maintainer}: Ultán P. Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul D. McNicholas \email{paulmc@mcmaster.ca} (\href{https://orcid.org/0000-0002-2482-523X}{ORCID})
  \item Arthur White \email{arwhite@tcd.ie} (\href{https://orcid.org/0000-0002-7268-5163}{ORCID})
}

}
\keyword{internal}
