% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierMBC-package.R
\docType{package}
\name{outlierMBC-package}
\alias{outlierMBC}
\alias{outlierMBC-package}
\title{outlierMBC: Iterative Detection & Identification of Outliers for Model-Based Clustering}
\description{
Exported Functions:
\itemize{
\item \code{ombc_gmm} - Identify outliers while clustering the data by fitting a
Gaussian mixture model.
\item \code{ombc_lcwm} - Identify outliers while performing regression and clustering
by fitting a linear cluster-weighted model to the data.
\item \code{simulate_gmm} - Simulate data from a Gaussian mixture model with outliers.
\item \code{simulate_lcwm} - Simulate data from a linear cluster-weighted model with
outliers.
\item \code{find_gross} -
\item \code{backtrack} -
\item \code{backtrack_gmm} -
\item \code{backtrack_lcwm} -
\item \code{plot_curve} -
\item \code{plot_backtrack} -
}
}
\author{
\strong{Maintainer}: Ultán Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
