% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{plot_backtrack}
\alias{plot_backtrack}
\title{Plot the outlier number selection curve for the backtrack method.}
\usage{
plot_backtrack(ombc_out, max_total_rise = 0.1, max_step_rise = 0.05)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}

\item{max_total_rise}{Upper limit for the cumulative increase, as a
proportion of the global minimum value, from all
backward steps.}

\item{max_step_rise}{Upper limit for the increase, as a proportion of the
global minimum value, from each backward step.}
}
\value{
A gg object.
}
\description{
Plot the outlier number selection curve for the backtrack method.
}
