% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtrack.R
\name{plot_backtrack_curve}
\alias{plot_backtrack_curve}
\title{Plot the outlier number selection curve for the backtrack method.}
\usage{
plot_backtrack_curve(ombc_out, max_total_rise = 0.1, max_step_rise = 0.01)
}
\arguments{
\item{ombc_out}{Output from ombc_gmm.}

\item{max_total_rise}{Value cannot exceed minimum * (1 + max_total_rise).}

\item{max_step_rise}{Each step must be less than minimum * (1 + max_step_rise).}
}
\value{
A gg object.
}
\description{
Plot the outlier number selection curve for the backtrack method.
}
