% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ombc.R
\name{plot_selection}
\alias{plot_selection}
\title{Plot dissimilarity values for multiple solutions.}
\usage{
plot_selection(ombc_list, x_axis = c("comp_num", "outlier_num", "param_num"))
}
\arguments{
\item{ombc_list}{A list of outputs from ombc_gmm.}

\item{x_axis}{The quantity to be plotted on the x axis.}
}
\value{
A gg object.
}
\description{
Plot dissimilarity values for multiple solutions.
}
