% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gmm.R
\name{simulate_gmm}
\alias{simulate_gmm}
\title{simulate_gmm}
\usage{
simulate_gmm(
  n,
  mu,
  sigma,
  outlier_num,
  seed = 123,
  crit_val = 0.9999,
  range_multiplier = 1.5,
  print_interval = Inf
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{range_multiplier}{How much greater should the range of the Uniform
samples be than the range of the Normal samples?}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
\code{data.frame} with columns:
\itemize{
\item $X1: first covariate
\item ...
\item $G: label
}
}
\description{
Simulate a Gaussian mixture model with multivariate outliers.
}
\examples{
gmm_k3n1000o10 <- simulate_gmm(
  n = c(500, 250, 250),
  mu = list(c(-1, 0), c(+1, -1), c(+1, +1)),
  sigma = list(diag(c(0.2, 4 * 0.2)), diag(c(0.2, 0.2)), diag(c(0.2, 0.2))),
  outlier_num = 10,
  seed = 123,
  crit_val = 0.9999,
  range_multiplier = 1.5
)

plot(
  gmm_k3n1000o10[, c("X1", "X2")],
  col = gmm_k3n1000o10$G + 1, pch = gmm_k3n1000o10$G + 1
)
}
