% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gmm.R
\name{simulate_gmm}
\alias{simulate_gmm}
\title{simulate_gmm}
\usage{
simulate_gmm(
  n,
  mu,
  sigma,
  outlier_num,
  seed = 123,
  crit_val = 0.9999,
  unif_range_multiplier = 1.5,
  print_interval = Inf
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{unif_range_multiplier}{How much greater should the range of the Uniform
samples be than the range of the Normal samples?}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
Matrix with a label column.
}
\description{
Simulate a Gaussian mixture model with multivariate outliers.
}
\examples{
n_vec <- c(2000, 1000, 1000)
mu_list <- list(c(-1, 0), c(+1, -1), c(+1, +1))
sigma_list <- list(
  diag(c(0.2, 4 * 0.2)),
  diag(c(0.2, 0.2)),
  diag(c(0.2, 0.2))
)
gmm_p2g3 <- simulate_gmm(
  n_vec, mu_list, sigma_list,
  outlier_num = 40, seed = 123, crit_val = 0.9999,
  unif_range_multiplier = 1.5
)
plot(gmm_p2g3[, 1:2],
  col = 1 + gmm_p2g3[, 3], pch = 1 + gmm_p2g3[, 3]
)
}
