% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{simulate_lcwm}
\alias{simulate_lcwm}
\title{simulate_lcwm}
\usage{
simulate_lcwm(
  n,
  mu,
  sigma,
  beta,
  error_sd,
  outlier_num,
  outlier_type = c("x_and_y", "x_only", "y_only"),
  seed = 123,
  prob_range = c(1e-08, 1e-06),
  range_multipliers = c(3, 3),
  print_interval = Inf,
  print_prob_g = FALSE
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{outlier_num}{Desired number of outliers.}

\item{outlier_type}{.}

\item{seed}{Seed.}

\item{prob_range}{Values for uniform sample rejection.}

\item{range_multipliers}{.}

\item{print_interval}{How frequently the iteration count is printed.}

\item{print_prob_g}{Print how likely a more extreme value would be for each
simulated outlier.}
}
\value{
\code{data.frame}:
\itemize{
\item $X1: first covariate
\item ...
\item $Y: response
\item $G: label
}
}
\description{
Simulate a multiple linear regression model with response variable outliers.
}
\examples{
lcwm_k3n1000o10 <- simulate_lcwm(
  n = c(300, 300, 400),
  mu = list(c(3), c(6), c(3)),
  sigma = list(as.matrix(1), as.matrix(0.1), as.matrix(1)),
  beta = list(c(0, 0), c(-75, 15), c(0, 5)),
  error_sd = c(1, 1, 1),
  outlier_num = c(3, 3, 4),
  outlier_type = "x_and_y",
  seed = 123,
  prob_range = c(1e-8, 1e-6),
  range_multipliers = c(1, 2)
)

plot(
  lcwm_k3n1000o10[, c("X1", "Y")],
  col = lcwm_k3n1000o10$G + 1,
  pch = lcwm_k3n1000o10$G + 1
)
}
