% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{simulate_lcwm}
\alias{simulate_lcwm}
\title{simulate_lcwm}
\usage{
simulate_lcwm(
  n,
  mu,
  sigma,
  beta,
  error_sd,
  outlier_num,
  outlier_type = c("x_and_y", "x_only", "y_only"),
  seed = 123,
  crit_val = 0.9999,
  range_multipliers = c(1.5, 1.5),
  print_interval = Inf
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{outlier_num}{Desired number of outliers.}

\item{outlier_type}{.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{range_multipliers}{.}

\item{print_interval}{How frequently the iteration count is printed.}
}
\value{
\code{data.frame}:
\itemize{
\item $X1: first covariate
\item ...
\item $Y: response
\item $G: label
}
}
\description{
Simulate a multiple linear regression model with response variable outliers.
}
\examples{
lcwm_k2n1000oxy10 <- simulate_lcwm(
  n = c(1000, 1000),
  mu = list(c(-1), c(+1)),
  sigma = list(as.matrix(0.2), as.matrix(0.2)),
  beta = list(c(1, 0), c(1, 3)),
  error_sd = c(1, 1),
  outlier_num = c(25, 25),
  outlier_type = "x_and_y",
  seed = 123,
  crit_val = 0.9999,
  range_multipliers = c(1.5, 2)
)

plot(
  lcwm_k2n1000oxy10[, c("X1", "Y")],
  col = lcwm_k2n1000oxy10$G + 1,
  pch = lcwm_k2n1000oxy10$G + 1
)
}
