% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_noisy_mlr.R
\name{simulate_noisy_mlr}
\alias{simulate_noisy_mlr}
\title{simulate_noisy_mlr}
\usage{
simulate_noisy_mlr(
  n,
  mu,
  sigma,
  beta,
  error_sd,
  outlier_num,
  seed = 123,
  crit_val = 0.9999,
  unif_range_multiplier = 1.5
)
}
\arguments{
\item{n}{Vector of component sizes.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{outlier_num}{Desired number of outliers.}

\item{seed}{Seed.}

\item{crit_val}{Critical value for uniform sample rejection.}

\item{unif_range_multiplier}{How much greater should the range of the Uniform
samples be than the range of the Normal samples?}
}
\value{
Matrix with a label column.
}
\description{
Simulate a multiple linear regression model with response variable outliers.
}
\examples{
n_vec <- c(1000)
mu_list <- list(+1)
sigma_list <- list(as.matrix(0.1))
beta_list <- list(c(1, 1))
error_sd_vec <- c(0.1)
noisy_mlr <- simulate_noisy_mlr(n_vec, mu_list, sigma_list, beta_list,
                                error_sd_vec,
                                outlier_num = 20, seed = 123,
                                crit_val = 0.999)
plot(x = noisy_mlr$covariates[, 1], y = noisy_mlr$responses,
     col = 1 + noisy_mlr$labels)
}
