% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_cwm.R
\name{test_outlier_ombc}
\alias{test_outlier_ombc}
\title{Check whether a new sample is an outlier for each component.}
\usage{
test_outlier_ombc(
  outlier_type,
  mu,
  sigma,
  beta,
  error_sd,
  x_sample,
  y_sample,
  crit_val
)
}
\arguments{
\item{outlier_type}{.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{x_sample}{New covariate sample.}

\item{y_sample}{New response sample.}

\item{crit_val}{Critical value for uniform sample rejection.}
}
\value{
Logical: TRUE if the new sample is an outlier for each component.
}
\description{
Check whether a new sample is an outlier for each component.
}
