% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lcwm.R
\name{uniform_outlier_ombc}
\alias{uniform_outlier_ombc}
\title{Produce a single sample that passes the outlier checks.}
\usage{
uniform_outlier_ombc(
  outlier_type,
  mu,
  sigma,
  beta,
  error_sd,
  g,
  uniform_spans,
  prob_range
)
}
\arguments{
\item{outlier_type}{.}

\item{mu}{List of component mean vectors.}

\item{sigma}{List of component covariance matrices.}

\item{beta}{List of component regression coefficient vectors.}

\item{error_sd}{Vector of component regression error standard deivations.}

\item{g}{Component index.}

\item{uniform_spans}{Covariate and response error spans.}

\item{prob_range}{Values for uniform sample rejection.}
}
\value{
Vector consisting of covariate values, response value, and label 0.
}
\description{
Produce a single sample that passes the outlier checks.
}
