% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ombc.R
\name{uniform_spans_lcwm}
\alias{uniform_spans_lcwm}
\title{Obtain the span of the observations for each component.}
\usage{
uniform_spans_lcwm(range_multipliers, covariates_g, errors_g)
}
\arguments{
\item{range_multipliers}{.}

\item{covariates_g}{Covariate values of the sampled observations.}

\item{errors_g}{Response errors of the sampled observations.}
}
\value{
\code{matrix}: minimum and maximum columns, rows for each covariate and
one for the response errors.
}
\description{
Obtain the span of the observations for each component.
}
